Implement scheduler + interruption
Currently, long running reads can block other readers when the following
pattern occurs:
- Long running read starts (acquire read lock)
- Write operation occurs (acquire write lock)
- Other read operation starts (blocks on acquiring read lock)
This is due to the implementation of Go's RWMutex where a pending
write blocks readers to ensure that writes are not starved.

To avoid long running readers blocking other readers this change
implements the following logic:
- Introduces a scheduler to replace the RWMutex that never blocks
  reads unless a write is in progress.
- Implements an interruption and retry mechanism that allows for long
  running reads to be interrupted and retried when writes have been
  pending for a significant amount of time

In addtion, this change moves some background work related to expiration
and AOF flushing to be protected behind a separate lock. This ensures
that frequent background work will not acquire write locks repeatedly
preventing the possibility of long running scans.